% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rail_periods.R
\docType{data}
\name{rail_periods}
\alias{rail_periods}
\title{Lookup table of four-week periods}
\format{A data frame with 568 rows and 5 variables:
\itemize{
\item \code{year}     The financial year (from 1 April to 31 May, e.g. "2011/12")
\item \code{period}   The number of the four-week period within the year (1 to 13)
\item \code{start}    The date of the first day of the period
\item \code{start}    The date of the last day of the period
\item \code{days}     The length of the four-week period in days.  Usually it is 28
days long, but some periods are shorter or longer so that each
year is covered by exactly 13 periods.
}}
\source{
The source is not disclosed.  It is similar to a
\href{https://www.whatdotheyknow.com/request/historic_knowledge_of_london_sta}{document}
obtained via the Freedom of Information Act, except the boundary between
periods 12 and 13 in the year 2009/10.  Another
\href{http://dataportal.orr.gov.uk/displayreport/report/html/d986b6bf-5ca7-45c4-800c-75156d93e1a8}{source}
by the Office for Rail and Road was used to confirm that boundary.  Other
sources have been seen but not used, for example
https://data.london.gov.uk/dataset/london-underground-performance-reports,
the file "tfl-tube-performance.xls".  Enquiries for definitive information
are being made through official channels.
}
\usage{
rail_periods
}
\description{
A lookup table of the four-week-ish 'Control Periods' used by the UK rail
}
\details{
Some series of the London Underground Services performance data are reported
in 'periods' approximately four weeks long.  Some periods are shorter or
longer so that each year is covered by exactly 13 periods, beginning on the
April the 1st.
}
\keyword{datasets}
