% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undergroundperformance-package.R
\docType{package}
\name{underground-package}
\alias{underground}
\alias{underground-package}
\title{underground: London Underground Performance Data}
\description{
Redistributes London Underground performance data in a more
convenient form for analysis than spreadsheets
}
\author{
\strong{Maintainer}: Duncan Garmonsway \email{duncan.garmonsway@gmail.com}

}
\keyword{internal}
