% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/underground.R
\docType{data}
\name{underground}
\alias{underground}
\title{London Underground Services Performance Data}
\format{A data frame with 108,482 rows and 9 variables:
\itemize{
\item \code{metric}   The title of the data series
\item \code{year}     Year beginning 1 April
\item \code{quarter}  The quarter, e.g. \code{"QTR 1"} of year beginning 1 April, for
quarterly series
\item \code{fourweek} The four-week period within the year beginning 1 April, for
four-weekly series. I don't know whether the 13th and final
period is a day longer.
\item \code{line}     The London Underground Line, for series broken down by line
\item \code{category} The cause of loss of customer hours, for some Lost Customer
Hours series
\item \code{asset}    Either \code{Escalators - MTBF (days)} or \code{Lifts - MTBF (days)}, for
the series "Escalator and Lift Mean Time Between Failures
(Days)"
\item \code{company}  The company concerned, for some Lost Customer Hours series (see
'Details')
\item \code{value}    The observed value of the metric
}}
\source{
https://tfl.gov.uk/corporate/publications-and-reports/underground-services-performance
}
\usage{
underground
}
\description{
A dataset of the London Underground services performance data timeseries from
2003/04 to 21 July 2018.
}
\details{
This dataset may be updated approximately monthly.  To use the latest data,
reinstall the package from the GitHub repository
https://github.com/nacnudus/underground.

There are many series, each given annually as well as one of quarterly and
four-weekly.  Most series are broken down by line, others are broken down by
other things as follows.
\itemize{
\item Various Lost Customer Hours series are broken down by \code{line}, \code{category}
(which is really 'cause'), and \code{company} (see 'Details')
\item Escalator and Lift Mean Time Between Failures (Days) is broken down by
\code{asset} (lift or escalator)
}

What the \code{company} column means:

\code{BCV} = Bakerloo, Central, Victoria (and Waterlool & City)
\code{SSL} = Sub-Surface Lines (District, Hammersmith and City, Metropolitan,
Circle and East London)
\code{JNP} = Jubilee, Northern, Picadilly

From https://en.wikipedia.org/wiki/Talk%3ATransport_for_London:

"The maintenance and upgrading of the London Underground Network is carried
out by different companies. Tube Lines in responsible for the Jubilee,
Northern and Piccadilly Lines (JNP), Metronet BCV is responsible for the
Bakerloo, Central and Victoria Lines (BCV) as well as the Waterloo and City
Line, whilst Metronet SSL (sub surface lines) is responsible for the District,
Hammersmith and City, Metropolitan, Circle and East London lines." DavidB601
20:57, 7 November 2006 (UTC)

The source spreadsheet contains a small dataset called  "Unmanned  - Open,
Stations" that is small and already tidy, so it isn't included in this
package.  It looks like this.

Month | Hours- Stations Open/Unmanned | Daily hours | Calendar Month Hours	| % Stations Open/Unmanned
-------|-------------------------------|-------------|-----------------------|-------------------------
Jan-18 |                        245.90 |     5240.00 |            162440.00	|                    0.15%
Feb-18 |                        142.23 |     5240.00 |            146720.00	|                    0.10%
Mar-18 |                        201.82 |     5240.00 |            162440.00	|                    0.12%
Apr-18 |                         90.4	|     5240.00 |            157200.00	|                    0.06%
May-18 |                        177.2  |     5240.00 |            162440.00	|                    0.11%
Jun-18 |                        229.43 |     5240.00 |            157200.00	|                    0.15%
Jul-18 |                        254.53 |     5240.00 |            162440.00	|                    0.16%
}
\keyword{datasets}
