% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/underground_colours.R
\docType{data}
\name{underground_colours}
\alias{underground_colours}
\alias{underground_colors}
\alias{underground_lines}
\title{London Underground Palette}
\format{An object of class \code{character} of length 13.}
\usage{
underground_colours

underground_colors

underground_lines
}
\description{
A palette for the London Underground lines, based on the official palette.
}
\details{
The same yellow as for the Circle Line is used for the combination of the
Circle and Hammersmith & City lines.

For 'All Lines', the pink colour of the Transport for London Visitor Centre
is used.

\code{underground_lines} is the complement of \code{underground_colours} (swapping the
names and values), for using as the \code{labels} argument to
ggplot2::scalel_*_identity().
}
\examples{
underground_colours
underground_colors
underground_lines

dontrun{
  underground \%>\%
    filter(metric == "Train delays longer than 15 minutes",
           year == "2016/17",
           is.na(fourweek),
           is.na(quarter),
           line != "All Lines") \%>\%
    mutate(fill= underground_colours[line]) \%>\%
    select(line, value, fill) \%>\%
    ggplot(aes(line, value, fill = fill)) +
    geom_bar(stat = "identity") +
    scale_fill_identity("", labels = underground_lines, guide = "legend") +
    theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
    xlab("") +
    ylab("") +
    ggtitle("Train delays longer than 15 minutes (2016/17)")
}
}
\keyword{datasets}
