% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rail_periods.R
\docType{data}
\name{rail_periods}
\alias{rail_periods}
\title{Lookup table of four-week periods}
\format{A data frame with 533 rows and 5 variables:
\itemize{
\item \code{year}     The financial year (from 1 April to 31 May, e.g. "2011/12")
\item \code{period}   The number of the four-week period within the year (1 to 13)
\item \code{start}    The date of the first day of the period
\item \code{start}    The date of the last day of the period
\item \code{days}     The length of the four-week period in days.  Usually it is 28
days long, but some periods are shorter or longer so that each
year is covered by exactly 13 periods.
}}
\source{
The data was obtained by private corresponce with the \href{orr.gov.uk}{Office for Road andRail}.  It is similar to data published
\href{http://dataportal.orr.gov.uk/displayreport/report/html/d986b6bf-5ca7-45c4-800c-75156d93e1a8}{online},
but over a much wider range of years.  Similar data published by others
differs slightly.  This package author regards the Office for Road and Rail
as the authoritative source.
}
\usage{
rail_periods
}
\description{
A lookup table of the four-week-ish 'Control Periods' used by the UK rail
}
\details{
Some series of the London Underground Services performance data are reported
in 'periods' approximately four weeks long.  Some periods are shorter or
longer so that each year is covered by exactly 13 periods, beginning on the
April the 1st.
}
\keyword{datasets}
